Split ParallelArrayCollection into separate library
Removes dependence of Parallel on Domain so that compiling
small executables (and low-level unit tests) doesn't require
building the entire code base (Domain pulls in many other
things like GR).
