+++ src/Parallel/ArrayCollection/CMakeLists.txt
@@ -1,13 +1,16 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-add_subdirectory(Tags)
+set(LIBRARY ParallelArrayCollection)
+
+add_spectre_library(${LIBRARY})
 
 spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
   CreateElementCollection.hpp
+  CreateElementsUsingDistribution.hpp
   DgElementArrayMember.hpp
   DgElementArrayMemberBase.hpp
   DgElementCollection.hpp
@@ -28,3 +31,19 @@ spectre_target_sources(
   PRIVATE
   DgElementArrayMemberBase.cpp
 )
+
+target_link_libraries(
+  ${LIBRARY}
+  PUBLIC
+  DataStructures
+  DomainStructure
+  ErrorHandling
+  Parallel
+  Serialization
+  Utilities
+  INTERFACE
+  Domain
+  DomainCreators
+  )
+
+add_subdirectory(Tags)
