+++ src/Parallel/ArrayCollection/CreateElementCollection.hpp
@@ -20,11 +20,11 @@
 #include "Domain/Tags/ElementDistribution.hpp"
 #include "Evolution/DiscontinuousGalerkin/Initialization/QuadratureTag.hpp"
 #include "Parallel/AlgorithmExecution.hpp"
+#include "Parallel/ArrayCollection/CreateElementsUsingDistribution.hpp"
 #include "Parallel/ArrayCollection/SpawnInitializeElementsInCollection.hpp"
 #include "Parallel/ArrayCollection/Tags/ElementCollection.hpp"
 #include "Parallel/ArrayCollection/Tags/ElementLocations.hpp"
 #include "Parallel/ArrayCollection/Tags/NumberOfElementsTerminated.hpp"
-#include "Parallel/CreateElementsUsingDistribution.hpp"
 #include "Parallel/GlobalCache.hpp"
 #include "Parallel/Info.hpp"
 #include "Parallel/Local.hpp"
